+++ app/assets/javascripts/folio/console/base.js
@@ -91,11 +91,12 @@
 //= require folio/console/files/batch/bar_component
 //= require folio/console/files/batch/form_component
 //= require folio/console/files/display_toggle_component
+//= require folio/console/files/index/modal_component
 //= require folio/console/files/picker/document_component
 //= require folio/console/files/picker/image_component
 //= require folio/console/files/picker_component
-//= require folio/console/files/show_component
 //= require folio/console/files/show/modal_component
+//= require folio/console/files/show_component
 //= require folio/console/flash/flash
 //= require folio/console/folio_attributes_fields_component
 //= require folio/console/form/errors/errors