+++ app/assets/javascripts/folio/modal.js
@@ -74,7 +74,6 @@ window.Folio.Stimulus.register('f-modal', class extends window.Stimulus.Controll
         autofocus.focus()
       }
 
-      this.dispatch('opened')
       this.element.dispatchEvent(new window.CustomEvent('f-modal:opened', { bubbles: true }))
     } else {
       this.unbindOutsideClick()
@@ -87,7 +86,6 @@ window.Folio.Stimulus.register('f-modal', class extends window.Stimulus.Controll
 
       this.removeBackdrops()
 
-      this.dispatch('closed')
       this.element.dispatchEvent(new window.CustomEvent('f-modal:closed', { bubbles: true }))
     }
   }