+++ app/components/folio/console/files/index_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Folio::Console::Files::IndexComponent < Folio::Console::ApplicationComponent
+  def initialize(file_klass:, files:, modal: false)
+    @file_klass = file_klass
+    @files = files
+    @modal = modal
+  end
+end