+++ app/helpers/folio/console/index_helper.rb
@@ -7,7 +7,7 @@ module Folio::Console::IndexHelper
     opts[:tabs] ||= index_tabs
     opts[:folio_console_merge] ||= @folio_console_merge
     opts[:csv] = controller.try(:folio_console_controller_for_handle_csv) if opts[:csv].nil?
-    cell("folio/console/index/header", @klass, opts).show.html_safe
+    cell("folio/console/index/header", opts[:klass] || @klass, opts).show.html_safe
   end
 
   def catalogue(records, options = {}, &block)